fix: stop profiling on main program exit. (#20)
* fix: stop profiling on main program exit.

To achieve this, we create a new watchdog thread that checks the
active thread count. When only the threads in the agent are active,
the main program has finished and we can stop profiling and exit.

This fixes #6.

* replace watchdog thread by a custom executor thread factory.

The executor thread factory creates a daemon thread for the executor,
which is automatically shutdown when the main thread ends.
This solution is simpler and works better.