feat: initial JFR format support.
Java Flight Recorder format is a binary format integrated into Java
and the only format in AsyncProfiler that supports multiple events
concurrently.

In order to support multiple concurrent events in the Java
integration, we have created a [JFR
parser](https://github.com/pyroscope-io/jfr-parser/) and [integrated
the parser](https://github.com/pyroscope-io/pyroscope/pull/954) in
Pyroscope.

Now, we are adding support to generate the profiling data in this
format in the agent. This is only a first iteration towards full,
multi-event JFR ingestion support but we are not there yet.

This first step just supports defining the format through a new
environment variable (PYROSCOPE_FORMAT). By default the
old (collapsed) format will be used, but setting the variable to jfr:

PYROSCOPE_FORMAT=jfr

will use the new JFR format instead.
Note that this is only supported for CPU events (cpu, itimer, wall),
the rest of events (alloc, locks) aren't currently support with this
format (for now).